+++ src/atmcirclib/geo.py
@@ -9,7 +9,7 @@ from typing import Optional
 from typing import Union
 
 # Third-party
-import cartopy.crs as ccrs
+import cartopy.crs as ccrs  # type: ignore [import]
 import numpy as np
 import numpy.typing as npt